+++ bocco-emo/bocco_tools.py
@@ -18,7 +18,7 @@ class BoccoTools:
         self._webhook_events_cb: Dict[str, Dict[str, Callable]] = {}
         self._request_id_deque: deque = deque([], 10)
         
-    # 音声認識
+    # 音声認識 (Webhookで代用できた。使わない可能性高)
     def get_speech(self) -> str:
         self.access_token, self.uuid = self.auth.check_expired()
         end_point = "/v1/rooms/"+self.uuid+"/messages"
@@ -61,7 +61,7 @@ class BoccoTools:
         print(f"POSTED: {body}")
     
     # webhookの設定
-    def set_webhook(self, webhook:Webhook) -> None:
+    def set_webhook(self, webhook:Webhook) -> Optional[WebhookInfo]:
         self.access_token, self.uuid = self.auth.check_expired()
         end_point = "/v1/webhook"
         header = {
@@ -73,22 +73,23 @@ class BoccoTools:
             "url": webhook.url
         }
         body = self.api.post(data=data, end_point=end_point, headers=header)
-        
+        print(body)
         if "error" in body:
-            return WebhookInfo(**body)
+            return 
+        return WebhookInfo(**body)
     
-    # event取得        
+    # eventに紐づいているCallback関数の取得（部屋のuuidごとに設定可能）        
     def event(self, event: str, room_id_list: List[str] = []) -> Callable:
         def decorator(func):
-            print(func)
+            if room_id_list == []:
+                room_id_list.append(self.uuid)
             if event not in self._webhook_events_cb:
                 self._webhook_events_cb[event] = {}
-
             for room_id in room_id_list:
                 self._webhook_events_cb[event][room_id] = func
-            print(self._webhook_events_cb)
         return decorator
     
+    # webhookにeventを登録する
     def _register_webhook_event(self, events: List[str]) -> dict:
         self.access_token, self.uuid = self.auth.check_expired()
         end_point = "/v1/webhook/events"
@@ -105,14 +106,15 @@ class BoccoTools:
             return {"error": body["error"]}
         return WebhookInfo(**body)
         
+    # Webhookでのevent監視を始める（上の関数（_register_webhook_event）を実行しているだけ, 戻り値が違う）
     def start_webhook_event(self) -> str:
         response = self._register_webhook_event(list(self._webhook_events_cb.keys()))
         return response.secret
 
+    # イベントが発火した時にCallback関数とイベントで取得したデータをクラスにして返す
     def get_cb_func(self, body: dict) -> Tuple[Callable, WebhookBody]:
         self.access_token, self.uuid = self.auth.check_expired()
         emo_webhook_body = WebhookBody(**body)
-        print(emo_webhook_body, self._webhook_events_cb)
         if emo_webhook_body.request_id not in self._request_id_deque:
             try:
                 event_cb = self._webhook_events_cb[emo_webhook_body.event]
@@ -127,7 +129,7 @@ class BoccoTools:
             else:
                 print(f"error: room error")
             self._request_id_deque.append(emo_webhook_body.request_id)
-            return cb_func
+            return cb_func, emo_webhook_body
         else:
             print("Webhook request id is duplicated")
