+++ bocco-emo/requirements.txt
@@ -7,7 +7,6 @@ attrs==24.2.0
 certifi==2024.6.2
 charset-normalizer==3.3.2
 distro==1.9.0
-emo-platform-api-sdk==1.0.3
 fire==0.4.0
 frozenlist==1.4.1
 h11==0.14.0
